pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cad-cli"
version = "0.1.0"
description = "Agent-native CLI: image → VLM-described geometry → parametric CAD"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "google-genai>=1.0.0",
    "cadquery>=2.4.0",
    "trimesh>=4.0.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
cad-cli = "cad_cli.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/cad_cli"]

[tool.hatch.build.targets.sdist]
include = ["src/cad_cli", "tests", "README.md", "pyproject.toml"]
